Refactor auth module to use Flask-JWT-Extended

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity
 from models.item import ItemModel
 from models.store import StoreModel
 
@@ -17,7 +17,7 @@ class Item(Resource):
         help="Every item needs a store id"
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
 
@@ -25,7 +25,7 @@ class Item(Resource):
             return item.json()
         return {'message': 'Item does not exist'}, 404
 
-    @jwt_required()
+    @jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': f'An item with name \'{name}\' already exists'}, 400
@@ -44,15 +44,20 @@ class Item(Resource):
         
         return item.json(), 201
 
-    @jwt_required()
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin access is required'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
+            return {'message': 'Item deleted'}
 
-        return {'message': 'Item deleted'}
+        return {'message': 'Item not found'}, 404
 
-    @jwt_required()
+    @jwt_required
     def put(self, name):
         data = Item.parser.parse_args()
 
@@ -72,6 +77,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
-    @jwt_required()
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'Product details are available only for logged in users'
+        }, 200
